mcp_servers/quote_mcp/pyproject.toml
[project]
name = "quote_mcp"
version = "0.1.0"
description = "MCP Server for generating furniture quotes"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.10"
